Implement python3 comparison protocol
py3 does not handle `__cmp__`.

Note further that python3 implicitly adds `__hash__=None` to the class
members during class construction if there is an `__eq__` defined. The
result is that we see errors like this:

    class X:
        def __hash__(self):
            return hash(self.val)
        def __eq__(self, other):
            return self.val == other.val

    >>> hash(conn.modules.__main__.X())
    TypeError: unhashable instance

To fix this we need to either have `__hash__` AND the comparison methods
in `_local_netref_attrs`or neither of them.

Fixes #280, #293, #267
Closes #281
Supersedes #268