Fix: default list shared across all Server objects
In python, the default values of function parameters are instantiated
at function definition time. All calls to that function that use the
default value all point to the same global object. e.g.:

```
def func(x=[]):
   x.append(1)
   print(x)

func() # [1]
func() # [1 , 1]
```

Because of this, two instances of Server (initialized without passing
in a protocol_config option) actually share the same protocol_config.
So modifying one server's config affects the other ones.

Fix:
The recommended solution is to either set default to None and assign
a new empty object when the variable is None.